python_server/requirements.txt
# --- Core FastAPI stack ---
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# --- PyTorch stack ---
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
setuptools==69.5.1
wheel==0.43.0

# --- Image processing ---
pillow==10.1.0
opencv-python==4.9.0.80
matplotlib==3.8.0

# --- Utils ---
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

# --- Optional (para seguridad y CORS) ---
starlette==0.27.0

# --- Tabular ML models ---
xgboost==1.7.6
scikit-learn==1.6.1
joblib==1.3.2
pandas==2.1.1